x86/mm: normalize the per-domain page counters on unsigned long
Per-domain page counts in struct domain (tot_pages, max_pages,
xenheap_pages, outstanding_pages, extra_pages) and in struct
paging_domain (total_pages, free_pages, p2m_pages) were stored as
unsigned int.

This caps each domain at 2^32 pages = 16TB and considering that
systems (and hence guests) only ever get larger, it likely should
all be unsigned long.

Changes in this commit:
- sched.h: tot_pages, xenheap_pages, outstanding_pages, max_pages
           and extra_pages become unsigned long;
           domain_tot_pages() returns unsigned long.
- x86/asm/domain.h: paging_domain.{total,free,p2m}_pages
           become unsigned long.
- assign_pages(): widen the local tot_pages copy
- xen/arch/x86/mm/shadow/common.c: widen shadow_set_allocation(pages)
- All %u/%i/PRIu32 print sites for the widened counters become %lu

Suggested-by: Jan Beulich <jbeulich@suse.com>
---
This commit is only the initial normalisation step and would need
quite a few more fixups to fix all truncations to unsigned int.

+++ xen/arch/x86/include/asm/domain.h
@@ -221,9 +221,9 @@ struct paging_domain {
 
     /* Memory allocation (common to shadow and HAP) */
     struct page_list_head   freelist;
-    unsigned int            total_pages;  /* number of pages allocated */
-    unsigned int            free_pages;   /* number of pages on freelists */
-    unsigned int            p2m_pages;    /* number of pages allocated to p2m */
+    unsigned long           total_pages;  /* number of pages allocated */
+    unsigned long           free_pages;   /* number of pages on freelists */
+    unsigned long           p2m_pages;    /* number of pages allocated to p2m */
 
     /* log dirty support */
     struct log_dirty_domain log_dirty;
